# fix: indexes to optimize principal-based etag db lookups